feature(lambda): Add lambda tests and client
