+++ tox.ini
@@ -20,6 +20,7 @@ envlist =
     py35-flask-12
     py35-flask-dev
     py27-celery-{3,4}
+    py{27-36}-lambda
 
 
 [testenv]
@@ -49,6 +50,7 @@ setenv =
     django: TESTPATH=tests/contrib/django
     flask: TESTPATH=tests/contrib/flask
     celery: TESTPATH=tests/contrib/test_celery.py
+    lambda: TESTPATH=tests/contrib/awslambda
 usedevelop = true
 extras = tests
 
@@ -58,6 +60,7 @@ basepython =
     py33: python3.3
     py34: python3.4
     py35: python3.5
+    py36: python3.6
     pypy: pypy
 
 commands =
